Made did some refactoring and broke out some functions. Think i fixed the bug by adding the storyblok dir to the project root when ran and removing it after
