+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mui-storyblok",
-  "version": "1.0.68",
+  "version": "1.0.69",
   "description": "use react and mui to build pages with storyblok",
   "author": "apjames93",
   "license": "MIT",
@@ -156,11 +156,12 @@
     "styleguide": "npx styleguidist server",
     "styleguide:build": "npx styleguidist build",
     "deploy-style-guide": "npm run styleguide:build && aws s3 sync styleguide/ s3://mui-storyblok-style-guide.com/ --delete",
-    "export-storyblok": "npx storyblok-component-export",
-    "test-report": "react-scripts test --watchAll=false --coverage"
+    "test-report": "react-scripts test --watchAll=false --coverage",
+    "export-storyblok-components": "npx storyblok-component-export",
+    "export-storyblok-stories": "npx storyblok-backup --stories"
   },
   "bin": {
-    "migrate-storyblok": "./cli.js"
+    "migrate-storyblok-components": "./cli/migrateStoryblokComponents.js"
   },
   "browserslist": {
     "production": [
@@ -185,7 +186,7 @@
     "react-player": "^2.1.0",
     "react-router-dom": "^4.3.1",
     "react-swipeable-views": "^0.13.9",
-    "rff-wrapper": "^1.0.2",
+    "rff-wrapper": "^1.0.12",
     "storyblok-migrate": "^0.3.3"
   }
 }
